pyproject.toml
[tool.poetry]
name = "py-mem-mcp"
version = "0.1.0"
description = "A Python implementation of the MEM ontology MCP server"
authors = ["vgoehler"]
license = "CC0-1.0"
readme = "README.md"
packages = [{include = "py_mem_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastmcp = "^2.0"
httpx = "^0.27"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"

[tool.poetry.scripts]
py-mem-mcp = "py_mem_mcp.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
